Refactor presets, session, selection & status
Major refactor of app behavior and utilities:

- Extract time formatting to gesturesesh.utils.time and use format_seconds where needed.
- Centralize config persistence by calling save_config from gesturesesh.update_checker instead of dynamic imports.
- Change preset storage to a wrapped format (dict with "schedule" and optional "session_settings"); save() updates existing wrapped presets and load() handles both legacy and new formats.
- Persist recent session display preferences (zoom, grayscale, resize, frameless, etc.) under recent_session and also store toggle_resize_status on the active preset so presets keep their own resize state.
- Simplify and clean up MainApp selection logic: use QFileDialog directly, streamline messages, remove legacy duplicate detection, image manager, and file-identity helpers.
- Simplify scan/load/save/remove logic across presets/selection/session mixins and remove duplicated code from main.py; main.py imports cleaned up accordingly.
- Rework status message system: simplify timers and blink/fade handling, make animations more robust to exceptions, adjust message lifetime behavior, and tidy HTML rendering.
- Add new gesturesesh/utils/time.py module (format_seconds) and update tests accordingly.

Overall these changes consolidate persistence, reduce duplicated code, improve robustness of UI status animations, and make preset/session display settings more consistent and maintainable.
