+++ src/gesturesesh/app/session.py
@@ -3,13 +3,13 @@
 from __future__ import annotations
 
 import os
-import importlib
 from pathlib import Path
 
 from PyQt5.QtTest import QTest
 
 from gesturesesh.app.models import ScheduleEntry
-from gesturesesh.session_window import BREAK_IMAGE_PATH
+from gesturesesh.session_window import BREAK_IMAGE_PATH, SessionDisplay
+from gesturesesh.update_checker import save_config
 
 
 class MainAppSessionMixin:
@@ -37,10 +37,9 @@ class MainAppSessionMixin:
         self.save(wait_status=False)
 
         self.insert_breaks()
-        main_module = importlib.import_module("gesturesesh.main")
-
-        # Pass any saved recent session settings into the display so initial
-        # state matches the user's last-used preferences.
+        # Pass recent session settings (if any) into the SessionDisplay so
+        # initial display state (resize, grayscale, zoom, etc.) is preserved.
+        session_settings = None
         try:
             session_settings = self.config.get("recent_session", {}).get(
                 "session_settings", None
@@ -48,7 +47,23 @@ class MainAppSessionMixin:
         except Exception:
             session_settings = None
 
-        self.display = main_module.SessionDisplay(
+        # toggle_resize_status is persisted per-preset rather than shared across
+        # all sessions, so override the global value with the active preset's
+        # value when present.
+        try:
+            preset_name = self.preset_loader_box.currentText()
+            preset = self.presets.get(preset_name)
+            if isinstance(preset, dict) and "session_settings" in preset:
+                preset_settings = preset.get("session_settings") or {}
+                if "toggle_resize_status" in preset_settings:
+                    session_settings = dict(session_settings or {})
+                    session_settings["toggle_resize_status"] = bool(
+                        preset_settings.get("toggle_resize_status")
+                    )
+        except Exception:
+            pass
+
+        self.display = SessionDisplay(
             schedule=self.session_schedule,
             items=self.selection["files"],
             total=self.total_scheduled_images,
@@ -58,19 +73,101 @@ class MainAppSessionMixin:
         self.display.show()
 
     def session_closed(self):
-        """Removes breaks, and displays status"""
+        """Removes breaks, displays status, and persists display preferences."""
         self.remove_breaks()
         self.display_status()
         self.activateWindow()
         self.raise_()
+        # Persist a compact snapshot of display-related toggles so subsequent
+        # sessions can restore the user's preferences (zoom, resize, grayscale,
+        # frameless, etc.). Stored under 'recent_session' -> 'session_settings'.
+        try:
+            if hasattr(self, "display") and self.display is not None:
+                ds = self.display
+                ss = {}
+                try:
+                    ss["zoom_enabled"] = bool(getattr(ds, "zoom_enabled", False))
+                except Exception:
+                    ss["zoom_enabled"] = False
+                try:
+                    ss["reset_zoom_between_images"] = bool(
+                        getattr(ds, "reset_zoom_between_images", True)
+                    )
+                except Exception:
+                    ss["reset_zoom_between_images"] = True
+                try:
+                    ss["default_zoom"] = float(getattr(ds, "default_zoom_factor", 1.0))
+                except Exception:
+                    ss["default_zoom"] = 1.0
+                try:
+                    imgmods = getattr(ds, "image_mods", {}) or {}
+                    ss["grayscale"] = bool(imgmods.get("grayscale", False))
+                    ss["grayscale_mode"] = imgmods.get(
+                        "grayscale_mode", imgmods.get("grayscale_mode", "perceptual")
+                    )
+                    ss["hflip"] = bool(imgmods.get("hflip", False))
+                    ss["vflip"] = bool(imgmods.get("vflip", False))
+                    try:
+                        ss["brightness"] = int(imgmods.get("brightness", 0))
+                    except Exception:
+                        ss["brightness"] = 0
+                    try:
+                        ss["contrast"] = float(imgmods.get("contrast", 1.0))
+                    except Exception:
+                        ss["contrast"] = 1.0
+                    ss["threshold"] = bool(imgmods.get("threshold", False))
+                    ss["edge"] = bool(imgmods.get("edge", False))
+                except Exception:
+                    ss["grayscale"] = False
+                    ss["grayscale_mode"] = "perceptual"
+                try:
+                    ss["toggle_resize_status"] = bool(
+                        getattr(ds, "toggle_resize_status", False)
+                    )
+                except Exception:
+                    ss["toggle_resize_status"] = False
+                try:
+                    ss["frameless_status"] = bool(getattr(ds, "frameless_status", False))
+                except Exception:
+                    ss["frameless_status"] = False
+                try:
+                    ss["always_on_top"] = bool(
+                        getattr(ds, "toggle_always_on_top_status", False)
+                    )
+                except Exception:
+                    ss["always_on_top"] = False
+
+                recent = self.config.get("recent_session", {})
+                recent["session_settings"] = ss
+                self.config["recent_session"] = recent
+
+                # Persist toggle_resize_status onto the active preset so each
+                # preset keeps its own resize configuration rather than
+                # inheriting the last session's value.
+                try:
+                    preset_name = self.preset_loader_box.currentText()
+                    if preset_name and preset_name in self.presets:
+                        preset = self.presets[preset_name]
+                        if not (isinstance(preset, dict) and "schedule" in preset):
+                            preset = {"schedule": preset}
+                        preset_settings = preset.get("session_settings") or {}
+                        preset_settings["toggle_resize_status"] = ss[
+                            "toggle_resize_status"
+                        ]
+                        preset["session_settings"] = preset_settings
+                        self.presets[preset_name] = preset
+                        self.config["presets"] = self.presets
+                except Exception:
+                    pass
+
+                save_config(self.config_path, self.config)
+        except Exception:
+            pass
+
         self.show_temporary_status("Recent session settings saved!", 3000)
 
     def is_valid_session(self):
-        """
-        Checks if all files exist, and
-        if there are enough images for the schedule.
-
-        """
+        """Checks if all files exist and there are enough images for the schedule."""
         for row in range(self.entry_table.rowCount()):
             items = []
             try:
@@ -93,7 +190,7 @@ class MainAppSessionMixin:
                 self.selection["files"].remove(file)
                 self.selected_items.setText(f"{os.path.basename(file)} not found!")
                 self.selected_items.append(
-                    f"Has the location or file name been changed?"
+                    "Has the location or file name been changed?"
                 )
                 self.selected_items.append(
                     "Image removed from selection."
@@ -114,7 +211,7 @@ class MainAppSessionMixin:
         return True
 
     def insert_breaks(self):
-        """Inserts break images as specified by the schedule"""
+        """Inserts break images as specified by the schedule."""
         if self.has_break:
             current_index = 0
             for entry in self.session_schedule:
@@ -125,15 +222,7 @@ class MainAppSessionMixin:
                     current_index += entry.images
 
     def remove_breaks(self):
-        """
-        Removes all occurrences of 'break.png' from the list of selected files.
-
-        Iterates through the 'files' list in reverse order and removes any file whose
-        basename is 'break.png'. This prevents issues with changing list indices during removal.
-
-        Returns:
-            None
-        """
+        """Removes all occurrences of break.png from the list of selected files."""
         i = len(self.selection["files"])
         while i > 0:
             i -= 1
@@ -141,7 +230,7 @@ class MainAppSessionMixin:
                 self.selection["files"].pop(i)
 
     def grab_schedule(self):
-        """Builds self.session_schedule with data from the schedule"""
+        """Builds self.session_schedule with data from the schedule."""
         self.session_schedule = []
         for row in range(self.entry_table.rowCount()):
             images = int(self.entry_table.item(row, 1).text())
@@ -151,26 +240,18 @@ class MainAppSessionMixin:
             self.session_schedule.append(ScheduleEntry(images, time))
 
     def save_to_recent(self):
-        """
-        Saves current session settings into unified config.json.
-        """
-        main_module = importlib.import_module("gesturesesh.main")
-
-        files_to_save = [
-            file_path
-            for file_path in self.selection["files"]
-            if file_path != BREAK_IMAGE_PATH
-        ]
+        """Saves current session settings into unified config.json."""
         previous = self.config.get("recent_session") or {}
         recent = {
-            "folders": list(self.selection["folders"]),
-            "files": list(files_to_save),
+            "folders": self.selection["folders"],
+            "files": self.selection["files"],
             "recent_preset": self.preset_loader_box.currentIndex(),
             "randomized": self.randomize_selection.isChecked(),
         }
-        # Preserve display preferences across runs; session_closed() refreshes
-        # them when the session window is closed.
+        # Preserve display preferences (resize, frameless, zoom, grayscale, etc.)
+        # so they survive across new sessions. They are refreshed on session
+        # close in session_closed().
         if "session_settings" in previous:
             recent["session_settings"] = previous["session_settings"]
         self.config["recent_session"] = recent
-        main_module.save_config(self.config_path, self.config)
+        save_config(self.config_path, self.config)
